Use `GCC` image for testing
We want to bump GCC to version 14 for running the tests.
However, GCC 14 would have to be installed from source.
It's not available as a package in the Python image (Debian bookworm) that we were using prior to this commit.
Installing from source is time-consuming and tedious.
This is why we're switching from the Python image to a GCC image (which includes Python by default).