Use a switch to lookup the IAT_ enum elements.
Fixes race condition bug: The lazy init was guarded with IAT_H defined, there is no guarentee that el_num was was zero initilized. Since the first element (IAT_H) is used to guard the array is not guarenteed to be completely initilized. The array could be declared all at once but it is no longer needed with the switch.