Replace memchr() check with a new function to check the element group.
In several places a memchr() is used to check which group an atom belongs to, e.g. is N,P,As, or Sb (nitrogen group) or O,S,Se,Te (oxygen group).
It is clearer to test this with a function with maps the integer down to the first element of the group, in some places there are checks for "nitrogen group but not nitrogen".